refactor: 추가 index.ts 삭제
