+++ src/views/register/index.ts
@@ -1,2 +1,4 @@
-export * from './model';
+export * from './model/types';
+export * from './model/usePostProjectLogo';
+export * from './model/usePostProjectRegistration';
 export { default as RegisterPage } from './ui/RegisterPage';
