--- src/views/register/model/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './usePostProjectLogo';
-export * from './usePostProjectRegistration';
