--- src/widgets/register/style/index.ts
@@ -1,1 +1,0 @@
-export * from './textStyle';
